Some console output:
```
master:
statMathStressTest took 11938999µs and 259609kB
statMathStressTest took 10276942µs and 210343kB
statMathStressTest took 10153385µs and 285419kB

this:
statMathStressTest took 5295211µs and 175893kB
statMathStressTest took 4379497µs and 173967kB
statMathStressTest took 4319839µs and 235630kB
```

The approach here is:
- Avoid creating new instances where possible
- No use of `toHashMap` within the class - left in so the rest of Unciv needs no changes
- No loops over the seven stats (but this is likely a minor part of the gains)

All remaining uses of toHashMap are more or less as iterator. Two more little helpers and every one of the remaining 21 calls would be much shorter and efficient - e.g. `.toHashMap().values.sum()` could become `.values.sum()`, `.toHashMap().maxByOrNull { it.value }!!.key` could be `.maxByOrNull { it.value }!!.key`, and so on. `toHashMap` would be gone.

Yes the test code is horrible - because it emulates current use patterns of the class to a degree. Improving the usage patterns would bring memory stress down a little further.

So - big question: Do we want to go down this road for a measly 110% performance gain of a single class?