@SomeTroglodyte 2021-08-11
I forget - there are several key components in Unciv already that depend on reflection, but which ones were those? I believe I even saw something within Gdx itself that does... I'll try to remember to holler when I next stumble upon one - or should I remember. Sad jokes aside, the all too ready availability of that `::` "gimme a boxed reference" operator speaks volumes I think. 

As I imagined it, the hashMap as iterator will be replaced by two sequences - one values only for aggregations and one as key/value pairs. They and all of the operations currently touching each field discretely could readily be rewritten into loops - but, squeeze every last cycle away...

The alternate approach would be to make the HashMap primary storage and to rewrite the discrete properties into using it as backing. Haven't measured that, I estimate it being on that 10sec to 4sec scale somewhere between 6 and 7, maybe 5. I'm lazy but holler loudly if you want that comparison. Oh, reflection-free. On the surface at least.

The actual implementation isn't the most important aspect here, but showing it can be done, that the potential gain is high, and the barely measurable gain due to relieving GC pressure (which in these runs most likely applied _after_ the run ended) is significant, too.