@yairm210 2021-08-12
One thing I would do is make the map be lazy initialized.
Most Stats aren't going to use that, and making a map IS more memory and cpu...