@SomeTroglodyte 2021-08-12
I had my doubts - lazy has _some_ overhead (I had the comparison on the road map anyway, I just didn't start out with it due to these prejudices), but that turned out to be an excellent idea. Final measurements, now including an implementation of the 'backing field _is_ already a map' idea:

current master: 10.4
pre-instantiated ::view: 4.6
lazy ::view: 3.8s :+1: :clap: :champagne: :partying_face: :birthday: 
map backing: 5.5s

I'll now clean up and document the thing - warning:
- toHashMap will stay but be reordered to the order of the enum. That changes the result of the `statMathRandomResultTest`, and potentially some toString() order. I want the order _consistent_. The function will be replaced anyway, and the replacement sequence _will_ have enum order.
- I'll reorder the methods a little more logically which will make the diff less readable. Functionally only the lazy bit and that ordering will change.