@SomeTroglodyte 2021-08-12
Since this is so basic, I'd like every active coder to take a look. @yairm210  could you invite them? I might be prejudiced and forget the most important ones...

FYI @All:
I'm done with phase 1 - this PR-, which was made under the constraint to produce a compatible Stats class without changing a single file in the rest of the core/* sources. Phase 2 will add a few API elements and remove `toHashMap` and subsequently touch a lot of client code - `Stats(gold = 1f)` instead of `Stats().apply { gold = 1f }`, `for ( (stat,value) in <Stats>` instead of using `toHashMap` for that, or `<Stats>.values.sum()` instead of the current construct...