@SomeTroglodyte 2021-08-15
> closing bracket

True. But doesn't change the outcome... And the follow-up is planned anyway.

>  not visible based on the name of the function

Very true. But - designed as drop-in, so no changing names. Yet. With name changes allowed - I'd say the non-inPlace would be the operator overrides only, so not times(), and either naming all inPlace with that suffix or some new naming idea. '`Mutant`'! (does that [no mutants allowed site](https://nma-fallout.com/) still exist?)